--- a/lib/fluent/config/types.rb
+++ b/lib/fluent/config/types.rb
@@ -71,7 +71,7 @@ module Fluent
       else
         # Current parser passes comment without actual values, e.g. "param #foo".
         # parser should pass empty string in this case but changing behaviour may break existing environment so keep parser behaviour. Just ignore comment value in boolean handling for now.
-        if str.respond_to?('start_with?') && str.start_with?('#')
+        if str.respond_to?(:start_with?) && str.start_with?('#')
           true
         elsif opts[:strict]
           raise Fluent::ConfigError, "#{name}: invalid bool value: #{str}"