--- a/lib/fluent/plugin.rb
+++ b/lib/fluent/plugin.rb
@@ -175,7 +175,7 @@ module Fluent
              else
                raise Fluent::ConfigError, "#{kind} plugin '#{type}' is not a Class nor callable (without arguments)."
              end
-      if parent && impl.respond_to?("owner=")
+      if parent && impl.respond_to?(:owner=)
         impl.owner = parent
       end
       impl.extend FeatureAvailabilityChecker